fix form error handling and refactor of the checkout page
