+++ pages/checkout/summary.tsx
@@ -1,31 +1,17 @@
-import { useContext, useEffect, useState } from 'react';
+import { useEffect, useState } from 'react';
 import { useRouter } from 'next/router';
 import Cookies from 'js-cookie';
-import NextLink from 'next/link';
 
-import {
-  Alert,
-  Link,
-  Box,
-  Button,
-  Card,
-  CardContent,
-  Divider,
-  Grid,
-  Typography,
-  Snackbar,
-} from '@mui/material';
+import { Grid, Typography } from '@mui/material';
 
 import { ShopLayout } from '../../components/layouts/ShopLayout';
-import { CartList, OrderSummary } from '../../components/cart';
-import { CartContext } from '@/context';
+import { CartList } from '../../components/cart';
+import { AlertErrorMessage } from '@/components/auth';
+import { SummaryInfo } from '@/components/checkout/SummaryInfo';
 
 const SummaryPage = () => {
   const router = useRouter();
-  const { shippingAddress, numberOfItems, createOrder } =
-    useContext(CartContext);
 
-  const [isPosting, setIsPosting] = useState(false);
   const [showError, setShowError] = useState(false);
   const [errorMessage, setErrorMessage] = useState('');
 
@@ -34,36 +20,6 @@ const SummaryPage = () => {
       router.push('/checkout/address');
     }
   }, [router]);
-
-  const onCreateOrder = async () => {
-    setIsPosting(true);
-
-    const { hasError, message } = await createOrder();
-    setShowError(hasError);
-
-    if (hasError) {
-      setIsPosting(false);
-      setErrorMessage(message);
-      return;
-    }
-
-    router.replace(`/orders/${message}`);
-  };
-
-  if (!shippingAddress) {
-    return <></>;
-  }
-
-  const {
-    firstName,
-    lastName,
-    address,
-    address2 = '',
-    city,
-    zip,
-    country,
-    phone,
-  } = shippingAddress;
 
   return (
     <ShopLayout
@@ -78,71 +34,12 @@ const SummaryPage = () => {
         <Grid item xs={12} sm={7}>
           <CartList />
         </Grid>
-        <Grid item xs={12} sm={5}>
-          <Card className="summary-card">
-            <CardContent>
-              <Typography variant="h2">
-                Resumen ({numberOfItems}{' '}
-                {numberOfItems === 1 ? 'producto' : 'productos'})
-              </Typography>
-              <Divider sx={{ my: 1 }} />
-
-              <Box display="flex" justifyContent="space-between">
-                <Typography variant="subtitle1">
-                  Dirección de entrega
-                </Typography>
-                <NextLink href="/checkout/address" passHref legacyBehavior>
-                  <Link underline="always">Editar</Link>
-                </NextLink>
-              </Box>
-
-              <Typography>
-                {firstName} {lastName}
-              </Typography>
-              <Typography>
-                {address}
-                {address2 ? `,${address2}` : ''}
-              </Typography>
-              <Typography>
-                {city}, {zip}
-              </Typography>
-              <Typography>{country}</Typography>
-              <Typography>{phone}</Typography>
-
-              <Divider sx={{ my: 1 }} />
-
-              <Box display="flex" justifyContent="end">
-                <NextLink href="/cart" passHref legacyBehavior>
-                  <Link underline="always">Editar</Link>
-                </NextLink>
-              </Box>
-
-              <OrderSummary />
-
-              <Box sx={{ mt: 3 }}>
-                <Button
-                  color="secondary"
-                  className="circular-btn"
-                  fullWidth
-                  onClick={onCreateOrder}
-                  disabled={isPosting}
-                >
-                  Confirmar Orden
-                </Button>
-              </Box>
-            </CardContent>
-          </Card>
-        </Grid>
+        <SummaryInfo
+          setErrorMessage={setErrorMessage}
+          setShowError={setShowError}
+        />
       </Grid>
-      <Snackbar
-        open={showError}
-        autoHideDuration={5000}
-        anchorOrigin={{ vertical: 'bottom', horizontal: 'right' }}
-      >
-        <Alert severity="error" sx={{ width: '100%' }} variant="filled">
-          {errorMessage}
-        </Alert>
-      </Snackbar>
+      <AlertErrorMessage errorMessage={errorMessage} showError={showError} />
     </ShopLayout>
   );
 };
